using the submit table and submut total calories for all storages
