+++ lib/breakfast.js
@@ -1,5 +1,7 @@
 "use strict"
 
+var global = require('./global-submit.js')
+
 class Breakfast {
   constructor(food=0, calories=0){
     this.food = food;
@@ -30,85 +32,22 @@ deleteBreakfast(event, that) {
       }
      }
    };
-    updateBreakfast(){
 
-     var breakfastCalorieTotal = 0
-     if (localStorage.getItem('breakfast') === null) {
-       var breakfast = "[]"
-       localStorage.setItem('breakfast', breakfast)
+updateBreakfast(storage, tableName){
+     var CalorieTotal = 0
+     if (localStorage.getItem(storage) === null) {
+       var array = "[]"
+       localStorage.setItem(storage, array)
      } else {
-       JSON.parse(localStorage.getItem('breakfast')).forEach(function(element, i){
-         submitBreakfast(element.food, element.calories, i);
-         breakfastCalorieTotal += parseInt(element.calories)
+       JSON.parse(localStorage.getItem(storage)).forEach(function(element, i){
+         global.submitTable(element.food, element.calories, i, tableName, storage);
+         CalorieTotal += parseInt(element.calories)
        })
-       submitBreakfastCalories(breakfastCalorieTotal)
+       global.submitTableCalories(CalorieTotal, tableName)
      }
    }
 }
+// global[name1](CalorieTotal)
 
-function submitBreakfast(food, calories, id){
-  var newRow = document.createElement('tr');
-  newRow.id = "breakfast-"+food+"-"+calories
-  var foodCell = document.createElement('td');
-  foodCell.innerText = food;
-  foodCell.id = "breakfast-food-td"
-  var calorieCell = document.createElement('td');
-  calorieCell.innerText = calories;
-  calorieCell.id = "breakfast-calories-td"
-  var deleteCell = document.createElement('td')
-  deleteCell.id = "breakfast-delete-"+food+"-"+calories
-  deleteCell.innerHTML = '<i id="trash" class="fa fa-trash-o" aria-hidden="true"></i>';
-
-  newRow.appendChild(foodCell);
-  newRow.appendChild(calorieCell);
-  newRow.appendChild(deleteCell);
-
-  $(newRow).data("id",id)
-
-  var breakfastTable = document.getElementById('breakfastTable')
-
-  breakfastTable.appendChild(newRow);
-}
-
-function submitBreakfastCalories(total=0) {
-  var newRow = document.createElement('tr');
-  newRow.id = "breakfast-calorie-total-tr"
-
-  var firstTd = document.createElement('td');
-  firstTd.innerHTML = "<strong>Total Calories</strong>"
-
-  var calorieTotalCell = document.createElement('td');
-  calorieTotalCell.innerText = total;
-  calorieTotalCell.id = "breakfast-calories-total-td"
-
-  var emptyTd = document.createElement('td');
-  emptyTd.innerText = ""
-
-  newRow.appendChild(firstTd);
-  newRow.appendChild(calorieTotalCell);
-  newRow.appendChild(emptyTd);
-
-  var breakfastTable = document.getElementById('breakfastTable')
-  breakfastTable.appendChild(newRow);
-  // remaining row
-  var remainingRow = document.createElement('tr');
-  remainingRow.id = "breakfast-calorie-total"
-
-  var firstRemainingTd = document.createElement('td');
-  firstRemainingTd.innerHTML = "<strong>Remaining Calories</strong>"
-
-  var calorieRemainingCell = document.createElement('td');
-  calorieRemainingCell.innerText = 400 - total;
-  calorieRemainingCell.id = "breakfast-remaining-total"
-
-  var emptyRemainingTd = document.createElement('td');
-  emptyRemainingTd.innerText = ""
-
-  remainingRow.appendChild(firstRemainingTd);
-  remainingRow.appendChild(calorieRemainingCell);
-  remainingRow.appendChild(emptyRemainingTd);
-
-  breakfastTable.appendChild(remainingRow);
-}
 
 module.exports = Breakfast
