+++ lib/index.js
@@ -114,7 +114,7 @@ function clearChecked(checkBox){
 
 function updateBreakfastStorage(){
   var breakfastClass = new Breakfast()
-  breakfastClass.updateBreakfast();
+  breakfastClass.updateBreakfast('breakfast', 'breakfastTable');
 }
 
 // lunch starts here
@@ -138,7 +138,7 @@ $('#lunchTable').on("click", ".fa-trash-o", function(event){
 
 function updateLunchStorage(){
   var lunchClass = new Lunch()
-  lunchClass.updateLunch();
+  lunchClass.updateLunch('lunch', 'lunchTable');
 }
 
 function lunchChecker(checkBox){
@@ -175,7 +175,7 @@ $('#dinnerTable').on("click", ".fa-trash-o", function(event){
 
 function updateDinnerStorage(){
   var dinnerClass = new Dinner()
-  dinnerClass.updateDinner();
+  dinnerClass.updateDinner('dinner', 'dinnerTable');
 }
 
 function dinnerChecker(checkBox){
